requirements.txt
Flask==0.12.2
Flask-JWT-Extended==3.3.4
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.6.2
psycopg2==2.7.3.2
SQLAlchemy==1.1.14
bcrypt==3.1.4
python-dotenv==0.7.1